feature : move client package from supabase to pkg
