+++ pkg/supabase/drivers/cloud/admin/user.go
@@ -7,7 +7,7 @@ import (
 	"strings"
 
 	"github.com/sev-2/raiden"
-	"github.com/sev-2/raiden/pkg/supabase/client/net"
+	"github.com/sev-2/raiden/pkg/client/net"
 	"github.com/sev-2/raiden/pkg/supabase/objects"
 )
