+++ pkg/supabase/drivers/local/meta/policy.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/lib/pq"
 	"github.com/sev-2/raiden"
-	"github.com/sev-2/raiden/pkg/supabase/client/net"
+	"github.com/sev-2/raiden/pkg/client/net"
 	"github.com/sev-2/raiden/pkg/supabase/objects"
 	"github.com/sev-2/raiden/pkg/supabase/query"
 	"github.com/sev-2/raiden/pkg/supabase/query/sql"
