+++ pkg/supabase/supabase.go
@@ -11,8 +11,8 @@ import (
 
 	"github.com/hashicorp/go-hclog"
 	"github.com/sev-2/raiden"
+	"github.com/sev-2/raiden/pkg/client/net"
 	"github.com/sev-2/raiden/pkg/logger"
-	"github.com/sev-2/raiden/pkg/supabase/client/net"
 	"github.com/sev-2/raiden/pkg/supabase/drivers/cloud"
 	"github.com/sev-2/raiden/pkg/supabase/drivers/cloud/admin"
 	"github.com/sev-2/raiden/pkg/supabase/drivers/local/meta"
