+++ rpc.go
@@ -11,8 +11,8 @@ import (
 	"regexp"
 	"strings"
 
+	"github.com/sev-2/raiden/pkg/client/net"
 	"github.com/sev-2/raiden/pkg/logger"
-	"github.com/sev-2/raiden/pkg/supabase/client/net"
 	"github.com/sev-2/raiden/pkg/utils"
 	"github.com/valyala/fasthttp"
 )
